% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgm-auxilaryfunctions.R
\name{bpr}
\alias{bpr}
\title{Bayesian Probit Regression (BPR)}
\usage{
bpr(y, X, offset = 0, theta, theta_0 = c(0, 0, 0), N_sim = 1)
}
\arguments{
\item{y}{Vector of binary responses.}

\item{X}{Matrix of predictors.}

\item{offset}{Optional offset for the linear predictor.}

\item{theta}{Initial values for the regression coefficients.}

\item{theta_0}{Prior mean for the regression coefficients.}

\item{N_sim}{Number of simulations to perform.}
}
\value{
A matrix of simulated values for the regression coefficients.
}
\description{
Performs Bayesian Probit Regression given the predictors and response.
}
