% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{rgm-package}
\alias{rgm-package}
\title{rgm: Advanced Inference with Random Graphical Models}
\description{
Implements state-of-the-art Random Graphical Models (RGMs) for multivariate data analysis across multiple environments, offering tools for exploring network interactions and structural relationships. Capabilities include joint inference across environments, integration of external covariates, and a Bayesian framework for uncertainty quantification. Applicable in various fields, including microbiome analysis. Methods based on Vinciotti, V., Wit, E., & Richter, F. (2023). "Random Graphical Model of Microbiome Interactions in Related Environments." \href{https://arxiv.org/abs/2304.01956}{arXiv:2304.01956}.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/franciscorichter/rgm/issues}
}

}
\author{
\strong{Maintainer}: Francisco Richter \email{richtf@usi.ch}

Other contributors:
\itemize{
  \item Veronica Vinciotti \email{veronica.vinciotti@unitn.it} [contributor]
  \item Ernst Wit \email{ernst.jan.camiel.wit@usi.ch} [contributor]
}

}
