% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgm-auxilaryfunctions.R
\name{rot}
\alias{rot}
\title{Rotate Locations}
\usage{
rot(loc)
}
\arguments{
\item{loc}{Matrix of locations to rotate.}
}
\value{
Matrix of rotated locations.
}
\description{
Rotates locations to align with the mean vector direction.
}
\examples{
# Example usage with a 2-column matrix representing locations.
loc <- matrix(rnorm(20), ncol = 2)
rotated_loc <- rot(loc)
}
