% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgm-auxilaryfunctions.R
\name{sample.data}
\alias{sample.data}
\title{Sample Data}
\usage{
sample.data(data, K, tpoints)
}
\arguments{
\item{data}{A list of matrices representing the data.}

\item{K}{A list of matrices representing the precision matrices for each data matrix in `data`.}

\item{tpoints}{A list containing two lists of matrices for lower and upper truncation points, respectively.}
}
\value{
A list of matrices with the sampled data.
}
\description{
This function generates sample data based on the provided parameters and truncation points.
}
